Address Gadi's review comments
- Inline endian.resolve() at use sites in load_value instead of
  extracting an upfront variable
- Update Version 18 docstring: remove backward-compat boilerplate,
  add mention of the endianness fix
- Consolidate v18 tests into a single test_v18.py with one class
  (TestV17VsV18); drop round-trips already covered by test_roundtrip.py
